[QL] Add `BTPayPalVault.enablePayPalAppSwitch` property (#1227)
* Add BTPayPalVault.enablePayPalAppSwitch property
* Add new BTPayPalVaultBaseRequest - this allows us to only use the new property in the PayPal Web flow and cannot be accessed in the PayPal Native Checkout module
