+++ Sources/BraintreePayPal/BTPayPalVaultRequest.swift
@@ -5,57 +5,39 @@ import BraintreeCore
 #endif
 
 ///  Options for the PayPal Vault flow.
-@objcMembers open class BTPayPalVaultRequest: BTPayPalRequest {
+@objcMembers public class BTPayPalVaultRequest: BTPayPalVaultBaseRequest {
 
     // MARK: - Public Properties
 
-    /// Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
-    public var offerCredit: Bool
-    
+    /// Optional: Used to determine if the customer will use the PayPal app switch flow.
+    /// Defaults to `false`.
+    /// - Note: This property is currently in beta and may change or be removed in future releases.
+    public var enablePayPalAppSwitch: Bool
+
     /// Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
     public var userAuthenticationEmail: String?
 
     // MARK: - Initializer
 
-    /// Initializes a PayPal Native Vault request
-    /// - Parameter offerCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
-    public init(offerCredit: Bool = false, userAuthenticationEmail: String? = nil) {
-        self.offerCredit = offerCredit
+    /// Initializes a PayPal Vault request
+    /// - Parameters:
+    ///   - offerCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
+    ///   - userAuthenticationEmail: Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
+    ///   - enablePayPalAppSwitch: Optional: Used to determine if the customer will use the PayPal app switch flow. Defaults to `false`.
+    ///   This property is currently in beta and may change or be removed in future releases.
+    public init(offerCredit: Bool = false, userAuthenticationEmail: String? = nil, enablePayPalAppSwitch: Bool = false) {
+        self.enablePayPalAppSwitch = enablePayPalAppSwitch
         self.userAuthenticationEmail = userAuthenticationEmail
-
-        super.init(hermesPath: "v1/paypal_hermes/setup_billing_agreement", paymentType: .vault)
+        super.init(offerCredit: offerCredit)
     }
 
-    // MARK: Public Methods
-
-    /// :nodoc: Exposed publicly for use by PayPal Native Checkout module. This method is not covered by semantic versioning.
-    @_documentation(visibility: private)
     public override func parameters(with configuration: BTConfiguration) -> [String: Any] {
-        let baseParameters = super.parameters(with: configuration)
-        var vaultParameters: [String: Any] = ["offer_paypal_credit": offerCredit]
+        var baseParameters = super.parameters(with: configuration)
 
-        if let billingAgreementDescription {
-            vaultParameters["description"] = billingAgreementDescription
-        }
-        
         if let userAuthenticationEmail {
-            vaultParameters["payer_email"] = userAuthenticationEmail
+            baseParameters["payer_email"] = userAuthenticationEmail
         }
 
-        if let shippingAddressOverride {
-            let shippingAddressParameters: [String: String?] = [
-                "line1": shippingAddressOverride.streetAddress,
-                "line2": shippingAddressOverride.extendedAddress,
-                "city": shippingAddressOverride.locality,
-                "state": shippingAddressOverride.region,
-                "postal_code": shippingAddressOverride.postalCode,
-                "country_code": shippingAddressOverride.countryCodeAlpha2,
-                "recipient_name": shippingAddressOverride.recipientName
-            ]
-
-            vaultParameters["shipping_address"] = shippingAddressParameters
-        }
-
-        return baseParameters.merging(vaultParameters) { $1 }
+        return baseParameters
     }
 }
